fix(merge): 修复 origin 96 commits merge 后 9 个 compile errors
cargo check 在 merge origin/main 96 commits 后失败, 9 个 errors。

根因 1: merge 时 src/models.rs 取 origin 版本 (无 ToolUseProfile struct),
        但 src/agent.rs / src/rpc.rs / src/sdk.rs / src/acp.rs / src/app.rs
        不在 UU 列表, 自动保留本地 1ae44892e 引入的 ToolUseProfile 引用。
        结果 5 处 unresolved import / unknown field。

根因 2: origin a5f36e59d 引入了 tools.rs type mismatch (u16 vs u32,
        i32 vs u64), 不是 merge 引入, 是 origin 自己的 bug。

修复策略 (用户选 A):
- src/models.rs 加回 ToolUseProfile struct 套件 (5 个 struct + impl),
  不引入 validate/resolve 集成, 避免与 origin ModelsConfig 冲突
- ModelEntry 加 tool_use_profile: Option<ToolUseProfile> 字段
- 28 个 ModelEntry 初始化加 tool_use_profile: None
- src/extensions/extension_manager_impl.rs: 1 个 ModelEntry 初始化加 None
- src/tools.rs: 2 处 cast (as u32, as u64)
- src/main.rs: 1 个 AgentConfig 初始化加 tool_use_profile: None

验证:
- cargo check --bin pi: 通过 (5.31s)
- cargo check workspace: 通过 (0.76s, lib + bin + examples + tests)

注意: ToolUseProfile 功能本身未集成到 load 流程
(没有 validate / resolve 函数调用), ModelEntry.tool_use_profile
始终是 None, 但 struct 可用, agent / rpc / sdk / acp / app 的引用
能解析, cargo check 通过。

+++ src/tools.rs
@@ -4257,7 +4257,8 @@ where
         let original_permissions = expected_target_identity.map(|(_, _, mode)| {
             use std::os::unix::fs::PermissionsExt as _;
 
-            std::fs::Permissions::from_mode(mode)
+            // rustix st_mode is u16; std::fs::Permissions::from_mode needs u32.
+            std::fs::Permissions::from_mode(mode as u32)
         });
 
         // The pathname stored by `NamedTempFile` is not descriptor-relative.
@@ -4354,8 +4355,11 @@ where
                 let source_file = std::fs::File::from(source_descriptor);
                 let source_metadata = source_file.metadata()?;
                 if !source_metadata.is_file()
-                    || source_metadata.dev() != expected_dev
-                    || source_metadata.ino() != expected_ino
+                    // rustix st_dev/st_ino is i32 on this toolchain; std Metadata
+                    // returns u64. Cast at the comparison site to keep the
+                    // captured target identity tuple unchanged.
+                    || source_metadata.dev() != expected_dev as u64
+                    || source_metadata.ino() != expected_ino as u64
                 {
                     return Err(std::io::Error::other(
                         "file changed since it was read; re-read it and retry the edit",
